Ope
So I made quite a few layout changes this morning in the game page. I created a new css file specifically for the game page cause there was tons of styles I felt only needed for that one page. I added a space for the stat outputs and such. I also added a potental blank space that can be used for something later down the line.